test/mpi: merge rma lock tests into a single source file
