+++ test/mpi/rma/Makefile.dtp
@@ -42,32 +42,32 @@ epochtest_SOURCES = epochtest.c
 getfence1_CPPFLAGS = $(AM_CPPFLAGS)
 getfence1_SOURCES = getfence1.c
 
-lock_contention_dt_CPPFLAGS = $(AM_CPPFLAGS)
-lock_contention_dt_SOURCES = lock_contention_dt.c
+lock_contention_dt_CPPFLAGS = -DMULTI_ORIGIN -DSINGLE_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lock_contention_dt_SOURCES = lock_x_dt.c
 
-lock_dt_CPPFLAGS = $(AM_CPPFLAGS)
-lock_dt_SOURCES = lock_dt.c
+lock_dt_CPPFLAGS = -DSINGLE_ORIGIN -DSINGLE_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lock_dt_SOURCES = lock_x_dt.c
 
-lock_dt_flush_CPPFLAGS = $(AM_CPPFLAGS)
-lock_dt_flush_SOURCES = lock_dt_flush.c
+lock_dt_flush_CPPFLAGS = -DSINGLE_ORIGIN -DSINGLE_TARGET -DUSE_FLUSH -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lock_dt_flush_SOURCES = lock_x_dt.c
 
-lock_dt_flushlocal_CPPFLAGS = $(AM_CPPFLAGS)
-lock_dt_flushlocal_SOURCES = lock_dt_flushlocal.c
+lock_dt_flushlocal_CPPFLAGS = -DSINGLE_ORIGIN -DSINGLE_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL $(AM_CPPFLAGS)
+lock_dt_flushlocal_SOURCES = lock_x_dt.c
 
-lockall_dt_CPPFLAGS = $(AM_CPPFLAGS)
-lockall_dt_SOURCES = lockall_dt.c
+lockall_dt_CPPFLAGS = -DSINGLE_ORIGIN -DMULTI_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lockall_dt_SOURCES = lock_x_dt.c
 
-lockall_dt_flush_CPPFLAGS = $(AM_CPPFLAGS)
-lockall_dt_flush_SOURCES = lockall_dt_flush.c
+lockall_dt_flush_CPPFLAGS = -DSINGLE_ORIGIN -DMULTI_TARGET -DUSE_FLUSH -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lockall_dt_flush_SOURCES = lock_x_dt.c
 
-lockall_dt_flushall_CPPFLAGS = $(AM_CPPFLAGS)
-lockall_dt_flushall_SOURCES = lockall_dt_flushall.c
+lockall_dt_flushall_CPPFLAGS = -DSINGLE_ORIGIN -DMULTI_TARGET -DUSE_FLUSH_ALL -DUSE_FLUSH_LOCAL_NONE $(AM_CPPFLAGS)
+lockall_dt_flushall_SOURCES = lock_x_dt.c
 
-lockall_dt_flushlocal_CPPFLAGS = $(AM_CPPFLAGS)
-lockall_dt_flushlocal_SOURCES = lockall_dt_flushlocal.c
+lockall_dt_flushlocal_CPPFLAGS = -DSINGLE_ORIGIN -DMULTI_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL $(AM_CPPFLAGS)
+lockall_dt_flushlocal_SOURCES = lock_x_dt.c
 
-lockall_dt_flushlocalall_CPPFLAGS = $(AM_CPPFLAGS)
-lockall_dt_flushlocalall_SOURCES = lockall_dt_flushlocalall.c
+lockall_dt_flushlocalall_CPPFLAGS = -DSINGLE_ORIGIN -DMULTI_TARGET -DUSE_FLUSH_NONE -DUSE_FLUSH_LOCAL_ALL $(AM_CPPFLAGS)
+lockall_dt_flushlocalall_SOURCES = lock_x_dt.c
 
 putfence1_CPPFLAGS = $(AM_CPPFLAGS)
 putfence1_SOURCES = putfence1.c
@@ -87,32 +87,32 @@ epochtest_struct_SOURCES = epochtest.c
 getfence1_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
 getfence1_struct_SOURCES = getfence1.c
 
-lock_contention_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lock_contention_dt_struct_SOURCES = lock_contention_dt.c
+lock_contention_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCK_CONTENTION_DT $(AM_CPPFLAGS)
+lock_contention_dt_struct_SOURCES = lock_x_dt.c
 
-lock_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lock_dt_struct_SOURCES = lock_dt.c
+lock_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCK_DT $(AM_CPPFLAGS)
+lock_dt_struct_SOURCES = lock_x_dt.c
 
-lock_dt_flush_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lock_dt_flush_struct_SOURCES = lock_dt_flush.c
+lock_dt_flush_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCK_DT_FLUSH $(AM_CPPFLAGS)
+lock_dt_flush_struct_SOURCES = lock_x_dt.c
 
-lock_dt_flushlocal_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lock_dt_flushlocal_struct_SOURCES = lock_dt_flushlocal.c
+lock_dt_flushlocal_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCK_DT_FLUSHLOCAL $(AM_CPPFLAGS)
+lock_dt_flushlocal_struct_SOURCES = lock_x_dt.c
 
-lockall_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lockall_dt_struct_SOURCES = lockall_dt.c
+lockall_dt_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCKALL_DT $(AM_CPPFLAGS)
+lockall_dt_struct_SOURCES = lock_x_dt.c
 
-lockall_dt_flush_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lockall_dt_flush_struct_SOURCES = lockall_dt_flush.c
+lockall_dt_flush_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCKALL_DT_FLUSH $(AM_CPPFLAGS)
+lockall_dt_flush_struct_SOURCES = lock_x_dt.c
 
-lockall_dt_flushall_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lockall_dt_flushall_struct_SOURCES = lockall_dt_flushall.c
+lockall_dt_flushall_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCKALL_DT_FLUSHALL $(AM_CPPFLAGS)
+lockall_dt_flushall_struct_SOURCES = lock_x_dt.c
 
-lockall_dt_flushlocal_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lockall_dt_flushlocal_struct_SOURCES = lockall_dt_flushlocal.c
+lockall_dt_flushlocal_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCKALL_DT_FLUSHLOCAL $(AM_CPPFLAGS)
+lockall_dt_flushlocal_struct_SOURCES = lock_x_dt.c
 
-lockall_dt_flushlocalall_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
-lockall_dt_flushlocalall_struct_SOURCES = lockall_dt_flushlocalall.c
+lockall_dt_flushlocalall_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT -DLOCKALL_DT_FLUSHLOCALALL $(AM_CPPFLAGS)
+lockall_dt_flushlocalall_struct_SOURCES = lock_x_dt.c
 
 putfence1_struct_CPPFLAGS = -DUSE_DTP_POOL_TYPE__STRUCT $(AM_CPPFLAGS)
 putfence1_struct_SOURCES = putfence1.c
